fix: image reszie lambda
