+++ .npmrc
@@ -1,0 +1,1 @@
+auto-install-peers = true
