+++ construct/layers/sharp-layer/nodejs/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@gys/sharp-layer",
+  "version": "1.0.0",
+  "dependencies": {
+    "sharp": "^0.31.3"
+  }
+}
