+++ construct/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "@vardario/cdk-nextjs-app",
+  "version": "0.1.8",
+  "description": "",
+  "main": "lib/index.js",
+  "files": [
+    "/lib"
+  ],
+  "scripts": {
+    "test": "vitest run",
+    "build": "tsc --build",
+    "watch": "tsc --build --watch"
+  },
+  "author": "Sahin Vardar",
+  "license": "MIT",
+  "peerDependencies": {
+    "@apimda/npm-layer-version": "^1.0.0",
+    "@aws-cdk/aws-apigatewayv2-alpha": "^2.67.0-alpha.0",
+    "@aws-cdk/aws-apigatewayv2-integrations-alpha": "^2.67.0-alpha.0",
+    "aws-cdk-lib": "^2.78.0",
+    "constructs": "^10.2.17"
+  },
+  "dependencies": {
+    "@aws-sdk/client-s3": "3.188.0",
+    "next": "^13.3.0",
+    "serverless-http": "^3.2.0"
+  },
+  "devDependencies": {
+    "@types/aws-lambda": "^8.10.119",
+    "@types/node": "^20.1.2",
+    "aws-lambda": "^1.0.7",
+    "node-mocks-http": "^1.12.2",
+    "sharp": "^0.32.4",
+    "testcontainers": "^10.0.1",
+    "typescript": "^5.0.4",
+    "vitest": "^0.34.1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:vardario/cdk-react-app.git"
+  },
+  "packageManager": "pnpm@8.3.1"
+}
