+++ example/nextjs-app/app/page.tsx
@@ -1,7 +1,13 @@
+import Image from "next/image";
+import image from "../assets/test.jpg";
+
 export default function Home() {
   return (
     <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      <h1 className="m-auto">Hello,World</h1>
+      <article className="m-auto">
+        <h1 className="text-center">Hello,World</h1>
+        <Image src={image} width={512} alt="An awesome image" />
+      </article>
     </main>
   );
 }
