+++ example/nextjs-app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "nextjs-app",
-  "version": "0.1.7",
+  "version": "0.1.8",
   "private": true,
   "scripts": {
     "dev": "next dev",
