--- tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "lib",
-    "incremental": true,
-    "lib": ["ES2021"],
-    "module": "commonjs",
-    "target": "ES2021",
-    "strict": true,
-    "resolveJsonModule": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "experimentalDecorators": true,
-    "declaration": true
-  },
-  "include": ["src/**/*"],
-}
